+++ CHANGES.md
@@ -1,7 +1,8 @@
 0.10 (unreleased)
 -----------------
 
-- No changes yet.
+- Use new 3D and flood fill subset state classes from glue to make storing
+  subsets much more efficient.
 
 0.9.2 (unreleased)
 ------------------