+++ CHANGES.md
@@ -7,6 +7,9 @@
 - Work around an issue on certain graphics cards which causes volume
   renderings to not appear correctly but instead of have stripe artifacts. [#303]
 
+- Improve performance for volume rendering for arrays larger than 2048
+  along one or more dimensions. [#303]
+
 0.9.2 (unreleased)
 ------------------